The rename: Int is long, Float is double, and the tree is lowercase
Step 3 of docs/TYPES.md, and the one that cannot be done by halves.
Every builtin type name in every Luce source in the repository moves
to its lowercase spelling — 19 `.luc` files, 200 site samples, the
executable specification, the compiler's own diagnostics — while `Int`
and `Float` still *mean* i64 and binary64.  It is therefore provably
meaning-preserving: benchmark outputs are byte-identical, the site's
200 samples verify unchanged, the C twins are untouched, and `std_spec`
did not move.

The TitleCase names are gone rather than kept as aliases, and a
program that writes one is told what it is called now, by name:

    the builtin types are lowercase: Int is written long

in both places a type name can stand — an annotation and a conversion
call.  `retiredSpelling` in `support/types.zig` is the table, and it
is checked *before* the arity rules, so `List(Int)` is told about
`List` rather than that it "takes no type arguments", which is a
sentence about a struct nobody declared.

**`tools/spelling.zig` is the guard the memo asks for by name**, and
it fails the build.  The rename was safe because the names were
aliases; the resize that follows is safe only if the rename was
*complete*, and a missed `array(Float, ...)` is the one migration
failure that type-checks — it becomes a 32-bit array and changes
results with no diagnostic anywhere.  So it greps the three places
Luce source lives: `.luc` files, fenced `luce` blocks in the site's
markdown, and multiline string literals in `specs/`.  A
`spelling:off`/`spelling:on` pair suspends it, for the one thing it is
for: the two specs whose whole subject is the refusal have to be able
to write the refused name, and a grep for `spelling:off` lists every
exemption in the tree.

**The audit begins**, with the one site whose failure at 32 bits would
be *silent*: `std/math.luc`'s five constants are annotated `double`.
Each carries twenty-one significant digits so it rounds to the nearest
binary64, and the module promises ~1e-14; at binary32 the epsilon is
1.19e-7 and the promise is fiction.  Every other unannotated site in
the corpus meets a `double` or a `long` in an expression, so the type
checker names each one at the resize rather than a large input naming
it in six months — which is why the rest of the audit belongs to that
step, where the compiler does the finding.

The mechanical passes were scoped so they could not touch a Zig
identifier: `.luc` whole-file, `.zig` only on multiline-string
continuations, then whole-file on the stages verified to mention these
words nowhere but in strings and prose.  `08_llvm/` was excluded by
hand throughout, because `Builder` there is `std.zig.llvm.Builder`.
Prose that the substitution bruised — `an Int` became `an long`, two
paragraphs opened lowercase — is repaired; `an array` keeps its `an`.

Eleven pinned diagnostic columns moved, because `Int` is four letters
now and `Float` six.  Each points at the same token it always did.

1131 tests pass; the site's 200 samples verify; benchmark outputs
byte-identical.

+++ src/luce/interpreter/test.zig
@@ -46,10 +46,10 @@ fn compileScript(source: []const u8) !mir.Program {
 test "interpreter memory is bounded by depth and data, not by calls made" {
     var program = try compileScript(
         \\struct Point:
-        \\    x: Int
-        \\    y: Int
+        \\    x: long
+        \\    y: long
         \\
-        \\func nudge(p: Point) -> Int:
+        \\func nudge(p: Point) -> long:
         \\    var scratch: Point
         \\    scratch.x = p.x + 1
         \\    return scratch.x + p.y
@@ -93,7 +93,7 @@ test "the explicit frame stack survives deep recursion" {
     // promise a different way — it counts frames — and `specs/` is
     // where the two are compared.
     var program = try compileScript(
-        \\func dive(left: Int) -> Int:
+        \\func dive(left: long) -> long:
         \\    if left == 0:
         \\        return 0
         \\    return dive(left - 1)
